+++ src/instamatic/_collections.py
